调整目录结构，所有模块间通信的数据结构放到protocol目录下

+++ src/trading_server/risk_management/common/fund_manager.cpp
@@ -48,8 +48,9 @@ void FundManager::OnOrderSent(const Order* order) {
     double changed = contract->size * order->req.volume * order->req.price * margin_rate;
     account_->cash -= changed;
     account_->frozen += changed;
-    spdlog::debug("Account: balance:{:.3f} frozen:{:.3f} margin:{:.3f}", account_->total_asset,
-                  account_->frozen, account_->margin);
+    spdlog::debug("Account: balance:{:.3f} frozen:{:.3f} margin:{:.3f}",
+                  static_cast<double>(account_->total_asset), static_cast<double>(account_->frozen),
+                  static_cast<double>(account_->margin));
   }
 }
 
@@ -73,8 +74,9 @@ void FundManager::OnOrderTraded(const Order* order, const Trade* trade) {
       account_->margin -= margin;
       account_->cash += margin;
       if (account_->margin < 0) account_->margin = 0;
-      spdlog::debug("Account: balance:{:.3f} frozen:{:.3f} margin:{:.3f}", account_->total_asset,
-                    account_->frozen, account_->margin);
+      spdlog::debug("Account: balance:{:.3f} frozen:{:.3f} margin:{:.3f}",
+                    static_cast<double>(account_->total_asset),
+                    static_cast<double>(account_->frozen), static_cast<double>(account_->margin));
     }
   } else if (trade->trade_type == TradeType::kCashSubstitution) {
     if (order->req.direction == Direction::kPurchase) {
@@ -96,8 +98,9 @@ void FundManager::OnOrderCanceled(const Order* order, int canceled) {
     account_->frozen -= changed;
     account_->cash += changed;
 
-    spdlog::debug("Account: balance:{:.3f} frozen:{:.3f} margin:{:.3f}", account_->total_asset,
-                  account_->frozen, account_->margin);
+    spdlog::debug("Account: balance:{:.3f} frozen:{:.3f} margin:{:.3f}",
+                  static_cast<double>(account_->total_asset), static_cast<double>(account_->frozen),
+                  static_cast<double>(account_->margin));
   }
 }
